認証設定（authOptions）をauth.tsという別ファイルに分離 route.tsファイルを簡素化し、Next.jsのApp Router規約に準拠
